# api/internal/docs/generated/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/fizzbuzz/:int1/:int2/:limit/:str1/:str2": {
            "get": {
                "description": "Executes the FizzBuzz algorithm based on the provided parameters.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FizzBuzz"
                ],
                "summary": "Run FizzBuzz",
                "operationId": "metrics.Counter =\u003e fizzbuzz.FizzBuzz",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "First integer to replace with str1",
                        "name": "int1",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Second integer to replace with str2",
                        "name": "int2",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The upper limit for the FizzBuzz sequence",
                        "name": "limit",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "String to replace multiples of int1",
                        "name": "str1",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "String to replace multiples of int2",
                        "name": "str2",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of strings representing the FizzBuzz sequence",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - invalid input parameters",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}